--- docs/header-includes.yaml
@@ -1,6 +1,0 @@
----
-header-includes: |
-     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.4.0/styles/atom-one-light.min.css">
-     <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.4.0/highlight.min.js"></script>
-     <script>hljs.highlightAll()</script>
-...
